Refactor NativeProxy on Android
