+++ android/src/main/cpp/NativeProxy.h
@@ -32,7 +32,7 @@ using namespace facebook::jni;
 class AnimationFrameCallback : public HybridClass<AnimationFrameCallback> {
  public:
   static auto constexpr kJavaDescriptor =
-      "Lcom/swmansion/reanimated/NativeProxy$AnimationFrameCallback;";
+      "Lcom/swmansion/reanimated/nativeProxy/AnimationFrameCallback;";
 
   void onAnimationFrame(double timestampMs) {
     callback_(timestampMs);
@@ -57,7 +57,7 @@ class AnimationFrameCallback : public HybridClass<AnimationFrameCallback> {
 class EventHandler : public HybridClass<EventHandler> {
  public:
   static auto constexpr kJavaDescriptor =
-      "Lcom/swmansion/reanimated/NativeProxy$EventHandler;";
+      "Lcom/swmansion/reanimated/nativeProxy/EventHandler;";
 
   void receiveEvent(
       jni::alias_ref<JString> eventKey,
@@ -87,7 +87,7 @@ class EventHandler : public HybridClass<EventHandler> {
 class SensorSetter : public HybridClass<SensorSetter> {
  public:
   static auto constexpr kJavaDescriptor =
-      "Lcom/swmansion/reanimated/NativeProxy$SensorSetter;";
+      "Lcom/swmansion/reanimated/nativeProxy/SensorSetter;";
 
   void sensorSetter(jni::alias_ref<JArrayFloat> value) {
     size_t size = value->size();
@@ -117,7 +117,7 @@ class SensorSetter : public HybridClass<SensorSetter> {
 class KeyboardEventDataUpdater : public HybridClass<KeyboardEventDataUpdater> {
  public:
   static auto constexpr kJavaDescriptor =
-      "Lcom/swmansion/reanimated/NativeProxy$KeyboardEventDataUpdater;";
+      "Lcom/swmansion/reanimated/nativeProxy/KeyboardEventDataUpdater;";
 
   void keyboardEventDataUpdater(int keyboardState, int height) {
     callback_(keyboardState, height);
