+++ android/src/main/java/com/swmansion/reanimated/Utils.java
@@ -8,7 +8,7 @@ import java.util.Map;
 
 public class Utils {
 
-  protected static boolean isChromeDebugger = false;
+  public static boolean isChromeDebugger = false;
 
   public static Map<String, Integer> processMapping(ReadableMap style) {
     ReadableMapKeySetIterator iter = style.keySetIterator();
