+++ android/src/main/java/com/swmansion/reanimated/nativeProxy/NativeProxyCommon.java
@@ -1,0 +1,174 @@
+package com.swmansion.reanimated.nativeProxy;
+
+import android.os.SystemClock;
+
+import com.facebook.proguard.annotations.DoNotStrip;
+import com.facebook.react.ReactApplication;
+import com.facebook.react.bridge.NativeModule;
+import com.facebook.react.bridge.ReactApplicationContext;
+import com.facebook.react.bridge.ReadableMap;
+import com.facebook.react.bridge.ReadableNativeArray;
+import com.facebook.react.devsupport.interfaces.DevSupportManager;
+import com.facebook.soloader.SoLoader;
+import com.swmansion.common.GestureHandlerStateManager;
+import com.swmansion.reanimated.NodesManager;
+import com.swmansion.reanimated.Scheduler;
+import com.swmansion.reanimated.keyboardObserver.ReanimatedKeyboardEventListener;
+import com.swmansion.reanimated.sensor.ReanimatedSensorContainer;
+import com.swmansion.reanimated.sensor.ReanimatedSensorType;
+
+import java.lang.ref.WeakReference;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
+
+public class NativeProxyCommon {
+    static {
+        SoLoader.loadLibrary("reanimated");
+    }
+
+    protected NodesManager mNodesManager;
+    protected final WeakReference<ReactApplicationContext> mContext;
+    protected Scheduler mScheduler;
+    private ReanimatedSensorContainer reanimatedSensorContainer;
+    private final GestureHandlerStateManager gestureHandlerStateManager;
+    private ReanimatedKeyboardEventListener reanimatedKeyboardEventListener;
+    private Long firstUptime = SystemClock.uptimeMillis();
+    private boolean slowAnimationsEnabled = false;
+
+    protected NativeProxyCommon(ReactApplicationContext context) {
+        mScheduler = new Scheduler(context);
+        mContext = new WeakReference<>(context);
+        reanimatedSensorContainer = new ReanimatedSensorContainer(mContext);
+        reanimatedKeyboardEventListener = new ReanimatedKeyboardEventListener(mContext);
+        addDevMenuOption();
+
+        GestureHandlerStateManager tempHandlerStateManager;
+        try {
+            Class<NativeModule> gestureHandlerModuleClass =
+                    (Class<NativeModule>)
+                            Class.forName("com.swmansion.gesturehandler.react.RNGestureHandlerModule");
+            tempHandlerStateManager =
+                    (GestureHandlerStateManager) context.getNativeModule(gestureHandlerModuleClass);
+        } catch (ClassCastException | ClassNotFoundException e) {
+            tempHandlerStateManager = null;
+        }
+        gestureHandlerStateManager = tempHandlerStateManager;
+    }
+
+    public Scheduler getScheduler() {
+        return mScheduler;
+    }
+
+    private void toggleSlowAnimations() {
+        slowAnimationsEnabled = !slowAnimationsEnabled;
+        if (slowAnimationsEnabled) {
+            firstUptime = SystemClock.uptimeMillis();
+        }
+    }
+
+    private void addDevMenuOption() {
+        // In Expo, `ApplicationContext` is not an instance of `ReactApplication`
+        if (mContext.get().getApplicationContext() instanceof ReactApplication) {
+            final DevSupportManager devSupportManager =
+                    ((ReactApplication) mContext.get().getApplicationContext())
+                            .getReactNativeHost()
+                            .getReactInstanceManager()
+                            .getDevSupportManager();
+
+            devSupportManager.addCustomDevOption(
+                    "Toggle slow animations (Reanimated)", this::toggleSlowAnimations);
+        }
+    }
+
+    @DoNotStrip
+    private void requestRender(AnimationFrameCallback callback) {
+        mNodesManager.postOnAnimation(callback);
+    }
+
+    @DoNotStrip
+    private void updateProps(int viewTag, Map<String, Object> props) {
+        mNodesManager.updateProps(viewTag, props);
+    }
+
+    @DoNotStrip
+    private void synchronouslyUpdateUIProps(int viewTag, ReadableMap uiProps) {
+        mNodesManager.synchronouslyUpdateUIProps(viewTag, uiProps);
+    }
+
+    @DoNotStrip
+    private String obtainProp(int viewTag, String propName) {
+        return mNodesManager.obtainProp(viewTag, propName);
+    }
+
+    @DoNotStrip
+    private void scrollTo(int viewTag, double x, double y, boolean animated) {
+        mNodesManager.scrollTo(viewTag, x, y, animated);
+    }
+
+    @DoNotStrip
+    private void setGestureState(int handlerTag, int newState) {
+        if (gestureHandlerStateManager != null) {
+            gestureHandlerStateManager.setGestureHandlerState(handlerTag, newState);
+        }
+    }
+
+    @DoNotStrip
+    private long getCurrentTime() {
+        if (slowAnimationsEnabled) {
+            final long ANIMATIONS_DRAG_FACTOR = 10;
+            return this.firstUptime + (SystemClock.uptimeMillis() - this.firstUptime) / ANIMATIONS_DRAG_FACTOR;
+        } else {
+            return SystemClock.uptimeMillis();
+        }
+    }
+
+    @DoNotStrip
+    private float[] measure(int viewTag) {
+        return mNodesManager.measure(viewTag);
+    }
+
+    @DoNotStrip
+    private void configureProps(ReadableNativeArray uiProps, ReadableNativeArray nativeProps) {
+        Set<String> uiPropsSet = convertProps(uiProps);
+        Set<String> nativePropsSet = convertProps(nativeProps);
+        mNodesManager.configureProps(uiPropsSet, nativePropsSet);
+    }
+
+    private Set<String> convertProps(ReadableNativeArray props) {
+        Set<String> propsSet = new HashSet<>();
+        ArrayList<Object> propsList = props.toArrayList();
+        for (int i = 0; i < propsList.size(); i++) {
+            propsSet.add((String) propsList.get(i));
+        }
+        return propsSet;
+    }
+
+    @DoNotStrip
+    private void registerEventHandler(EventHandler handler) {
+        handler.mCustomEventNamesResolver = mNodesManager.getEventNameResolver();
+        mNodesManager.registerEventHandler(handler);
+    }
+
+    @DoNotStrip
+    private int registerSensor(int sensorType, int interval, SensorSetter setter) {
+        return reanimatedSensorContainer.registerSensor(
+                ReanimatedSensorType.getInstanceById(sensorType), interval, setter);
+    }
+
+    @DoNotStrip
+    private void unregisterSensor(int sensorId) {
+        reanimatedSensorContainer.unregisterSensor(sensorId);
+    }
+
+    @DoNotStrip
+    private int subscribeForKeyboardEvents(KeyboardEventDataUpdater keyboardEventDataUpdater, boolean isStatusBarTranslucent) {
+        return reanimatedKeyboardEventListener.subscribeForKeyboardEvents(keyboardEventDataUpdater, isStatusBarTranslucent);
+    }
+
+    @DoNotStrip
+    private void unsubscribeFromKeyboardEvents(int listenerId) {
+        reanimatedKeyboardEventListener.unsubscribeFromKeyboardEvents(listenerId);
+    }
+}
